divided refinement files into diff files for easier maintainability
